+++ src/transceivercontroller/TransceiverController.ts
@@ -65,7 +65,7 @@ export default interface TransceiverController {
   /**
    * Sets video sending bitrate in Kilo-bit-per-second
    */
-  setVideoSendingBitrateKbps(bitrateKbps: number, scaleResolutionDownBy?: number): void;
+  setVideoSendingBitrateKbps(bitrateKbps: number): void;
 
   /**
    * Returns the [[RTCRtpTransceiver]] for audio
@@ -78,7 +78,17 @@ export default interface TransceiverController {
   localVideoTransceiver(): RTCRtpTransceiver;
 
   /**
-   * Set [[RTCRtpEncodingParameters]] on the sender of transceiver
+   * Set [[RTCRtpEncodingParameters]] on the sender of transceiver.
+   * This method should be called whenever the sender's encoding parameters of the local video transceiver need to
+   * be updated.
+   * For example, the default NScaleVideoUplinkBandwidthPolicy calls this method whenever a video is on/off or the
+   * active speaker changes.
+   * This method assumes that the sender of the local video transceiver is available and the input parameters should
+   * not be empty.
+   * The encoding parameters for sender should be retrieved using sender.getParameters and updated using
+   * sender.setParameters method.
+   * @param {Map<string, RTCRtpEncodingParameters>} params - The encoding parameters. If you have multiple encoding
+   * parameters for different video layers, the key should be the rid corresponding to the RTCRtpEncodingParameters.
    */
   setEncodingParameters(params: Map<string, RTCRtpEncodingParameters>): void;
 }